requirements.txt
fastapi~=0.115
loguru~=0.7
pathlib~=1.0
setuptools~=75.8
uvicorn~=0.34
pydantic-settings~=2.9
PyYAML>=6.0
bcrypt~=4.1.2
pyjwt~=2.8.0
cachetools~=5.3.3
strawberry-graphql~=0.225.0